+++ NEWS.md
@@ -17,6 +17,7 @@
 * added function observation_matrix.
 * added function reward_matrix.
 * POMDP model now also contains horizon and terminal_values.
+* added MDP formulated as a POMDP.
 
 # pomdp 0.9.2 (12/06/2019)